AngularSpectrumMethod

Documentation for AngularSpectrumMethod.

AngularSpectrumMethod.ASM

AngularSpectrumMethod.BandLimitedASM

AngularSpectrumMethod.ScaledASM

AngularSpectrumMethod.ShiftedASM

AngularSpectrumMethod.TiltedASM

AngularSpectrumMethod.ASM — Method

ASM(u, λ, Δx, Δy, z; expand=true)

return diffracted field by the angular spectrum method (ASM). Evanescent waves are not eliminated but attenuated as $\exp(-2{\pi}wz)$. Without attenuation, the total energy $\iint|u|\mathrm{d}x\mathrm{d}y$ is conserved.

Arguments

u: input field.

λ: wavelength.

Δx: sampling interval the in x-axis.

Δy: sampling interval the in y-axis.

z: diffraction distance.

expand=true: if true (default), perform 4× expansion and zero padding for aliasing suppression.

Note

The x-axis is the horizontal direction, and the y-axis is the vertical.

AngularSpectrumMethod.BandLimitedASM — Method

BandLimitedASM(u, λ, Δx, Δy, z; expand=true)

return diffracted field by the band-limited ASM (see Ref. 1).

Kyoji Matsushima and Tomoyoshi Shimobaba, "Band-Limited Angular Spectrum Method for Numerical Simulation of Free-Space Propagation in Far and Near Fields," Opt. Express 17, 19662-19673 (2009)

AngularSpectrumMethod.ScaledASM — Method

ScaledASM(u, λ, Δx, Δy, z, R; expand=true)

return scaled diffraction field according to the scale factor $R$ by the scaled ASM (see Ref. 1).

Tomoyoshi Shimobaba, Kyoji Matsushima, Takashi Kakue, Nobuyuki Masuda, and Tomoyoshi Ito, "Scaled angular spectrum method," Opt. Lett. 37, 4128-4130 (2012)

AngularSpectrumMethod.ShiftedASM — Method

ShiftedASM(u, λ, Δx, Δy, z, x₀, y₀; expand=true)

return shifted diffraction field with the shift distance $x_{0}$ and $y_{0}$ by the shifted ASM (see Ref. 1).

Kyoji Matsushima, "Shifted angular spectrum method for off-axis numerical propagation," Opt. Express 18, 18453-18463 (2010)

AngularSpectrumMethod.TiltedASM — Method

TiltedASM(u, λ, Δx, Δy, T; expand=true, weight=false)

return tilted diffraction field for a rotation matrix $T$ by the tilted ASM (see Ref. 1, 2). If weight=true, a diagonal weighting matrix is used as the Jacobian determinant (default false). In this case, the energy conservation improves, but the computational cost is high (see Ref. 3).

Note

Rotations.jl is helpful in generating rotation matrices.

Kyoji Matsushima, Hagen Schimmel, and Frank Wyrowski, "Fast calculation method for optical diffraction on tilted planes by use of the angular spectrum of plane waves," J. Opt. Soc. Am. A 20, 1755-1762 (2003)

Kyoji Matsushima, "Formulation of the rotational transformation of wave fields and their application to digital holography," Appl. Opt. 47, D110-D116 (2008)

James G. Pipe and Padmanabhan Menon, "Sampling density compensation in MRI: Rationale and an iterative numerical solution," Magn. Reson. Med. 41, 179-186 (1999)